Refactor EEPROM access logic, limits ROM range when EEPROM is used
- Introduce eepromSetSize() to set proper eeprom size and address mask.
- When EEPROM is enabled and in use, access outside the usable ROM size
  region are ignored instead of running rom out-of-bounds logic
  - For 32 MiB ROMs, usable reads are limited to 0x01FFFF00.
  - for 16 MiB or smaller ROMs, access beyond 0x01000000 are ignored
    rather than triggering rom out-of-bounds logic (fixes some variant of Diddy Kong Pilot (prototype)).
  - Move back CPUReadROMxx logic back to CPUReadMemoryxx
  - Add helper for rom out-of-bounds logic
  - Optimize find_string()
  - Cleanedup various gba save logic on init (libretro)
  - other minor fixes

+++ src/core/gba/gbaSound.cpp
@@ -341,7 +341,7 @@ static void end_frame(blip_time_t time)
 #ifdef __LIBRETRO__
 void flush_samples(Multi_Buffer* buffer)
 {
-    int numSamples = buffer->samples_avail();
+    size_t numSamples = buffer->samples_avail();
     if ((numSamples * 2) > sizeof(soundFinalWave)) {
         // when there is more than 1-frame worth of samples, we
         // split it so as not to overflow our buffer.
